mv, img simulation to separate folder
