+++ deeplens/geolens.py
@@ -50,7 +50,7 @@ from deeplens.optics.geometric_surface import (
 )
 from deeplens.optics.phase_surface import Phase
 from deeplens.optics.material import Material
-from deeplens.optics.monte_carlo import forward_integral
+from deeplens.optics.imgsim import forward_integral
 from deeplens.optics.light import Ray
 from deeplens.optics.utils import diff_float
 from deeplens.optics.light import AngularSpectrumMethod
